Trains ML model to align images to ideal image
This machine learning script will discover unique features amongst the data set and use them to align to an ideal image (if available, must be labeled _ideal).  And then it will give training and vaildation reports as well as a list of the moves needed to align every image to the ideal image.